+++ src/apps/companies/apps/investments/large-capital-profile/client/state.js
@@ -1,5 +1,2 @@
-export const TASK_SAVE_LARGE_CAPITAL_INVESTOR_DETAILS =
-  'TASK_SAVE_LARGE_CAPITAL_INVESTOR_DETAILS'
-
 export const TASK_SAVE_LARGE_CAPITAL_INVESTOR_LOCATION =
   'TASK_SAVE_LARGE_CAPITAL_INVESTOR_LOCATION'
