+++ src/client/components/Form/elements/FieldCurrency/index.jsx
@@ -81,6 +81,7 @@ const FieldCurrency = ({
   initialValue,
   reduced,
   boldLabel,
+  currencySymbol = '£',
   ...rest
 }) => {
   const { value, error, touched, onChange, onBlur } = useField({
@@ -98,7 +99,7 @@ const FieldCurrency = ({
             error={touched && Boolean(error)}
             aria-hidden="true"
           >
-            £
+            {currencySymbol}
           </StyledCurrencyPrefix>
           <Input
             key={name}
@@ -159,6 +160,10 @@ FieldCurrency.propTypes = {
    * Boolean for rendering the label in bold or not
    */
   boldLabel: PropTypes.bool,
+  /**
+   * Sets the value for the currency prefix
+   */
+  currencySymbol: PropTypes.string,
 }
 
 FieldCurrency.defaultProps = {
