+++ src/client/components/Resource/InvestorTypes.js
@@ -1,0 +1,6 @@
+import { createMetadataResource } from '.'
+
+export default createMetadataResource(
+  'InvestorType',
+  'capital-investment/investor-type'
+)
