+++ src/client/index.jsx
@@ -24,7 +24,6 @@ import CreateListForm from '../apps/company-lists/client/CreateListForm'
 import DnbHierarchy from '../apps/companies/apps/dnb-hierarchy/client/DnbHierarchy'
 import LeadAdvisers from '../apps/companies/apps/advisers/client/LeadAdvisers'
 import LargeCapitalProfileCollection from '../apps/investments/client/profiles/LargeCapitalProfileCollection'
-import EditLargeCapitalInvestorDetails from '../apps/companies/apps/investments/large-capital-profile/client/EditLargeCapitalInvestorDetails'
 import EditLargeCapitalInvestorLocation from '../apps/companies/apps/investments/large-capital-profile/client/EditLargeCapitalInvestorLocation'
 import UnfilteredLargeCapitalOpportunityCollection from '../apps/investments/client/opportunities/List/UnfilteredLargeCapitalOpportunityCollection'
 import InvestmentEditHistory from '../apps/investments/client/InvestmentEditHistory'
@@ -218,11 +217,11 @@ import {
 } from '../apps/investments/client/projects/create/state'
 import * as createInvestmentProjectTasks from '../apps/investments/client/projects/create/tasks'
 
-import {
-  TASK_SAVE_LARGE_CAPITAL_INVESTOR_DETAILS,
-  TASK_SAVE_LARGE_CAPITAL_INVESTOR_LOCATION,
-} from '../apps/companies/apps/investments/large-capital-profile/client/state'
+import { TASK_SAVE_LARGE_CAPITAL_INVESTOR_LOCATION } from '../apps/companies/apps/investments/large-capital-profile/client/state'
 import * as updateLargeCapitalInvestorDetails from '../apps/companies/apps/investments/large-capital-profile/client/tasks'
+
+import { TASK_SAVE_LARGE_CAPITAL_INVESTOR_DETAILS } from './modules/Companies/CompanyInvestments/LargeCapitalProfile/state'
+import { updateInvestorDetails } from './modules/Companies/CompanyInvestments/LargeCapitalProfile/tasks'
 
 import {
   TASK_SAVE_ORDER_ASSIGNEES,
@@ -469,8 +468,7 @@ function App() {
             overviewInvestmentProjectTasks.getLatestExportWins,
           [TASK_CREATE_INVESTMENT_PROJECT]:
             createInvestmentProjectTasks.createInvestmentProject,
-          [TASK_SAVE_LARGE_CAPITAL_INVESTOR_DETAILS]:
-            updateLargeCapitalInvestorDetails.updateInvestorDetails,
+          [TASK_SAVE_LARGE_CAPITAL_INVESTOR_DETAILS]: updateInvestorDetails,
           [TASK_SAVE_LARGE_CAPITAL_INVESTOR_LOCATION]:
             updateLargeCapitalInvestorDetails.updateInvestorLocation,
           [TASK_GET_INVESTMENT_PROJECT_INITIAL_VALUES]:
@@ -679,9 +677,6 @@ function App() {
         </Mount>
         <Mount selector="#large-capital-profile-collection">
           {(props) => <LargeCapitalProfileCollection {...props} />}
-        </Mount>
-        <Mount selector="#edit-large-capital-investor-details">
-          {(props) => <EditLargeCapitalInvestorDetails {...props} />}
         </Mount>
         <Mount selector="#edit-large-capital-investor-location">
           {(props) => <EditLargeCapitalInvestorLocation {...props} />}
