+++ src/client/modules/Companies/CompanyInvestments/LargeCapitalProfile/constants.js
@@ -1,0 +1,7 @@
+export const requiredCheckTypes = {
+  CLEARED: 'Cleared',
+  ISSUES_IDENTIFIED: 'Issues identified',
+  NOT_YET_CHECKED: 'Not yet checked',
+  CHECKS_NOT_REQUIRED:
+    'Checks not required - See Investor Screening Report (ISR) guidance',
+}
