+++ src/client/modules/Companies/CompanyInvestments/LargeCapitalProfile/state.js
@@ -1,0 +1,2 @@
+export const TASK_SAVE_LARGE_CAPITAL_INVESTOR_DETAILS =
+  'TASK_SAVE_LARGE_CAPITAL_INVESTOR_DETAILS'
