+++ src/client/modules/Companies/CompanyInvestments/LargeCapitalProfile/tasks.js
@@ -1,6 +1,6 @@
-import { apiProxyAxios } from '../../../../../../client/components/Task/utils'
+import { apiProxyAxios } from '../../../../components/Task/utils'
 
-export const updateInvestorLocation = (values) => {
+export const updateInvestorDetails = (values) => {
   const { id, investor_company_id } = values
   return apiProxyAxios.patch(
     `v4/large-investor-profile/${id}?investor_company_id=${investor_company_id}`,
